Fix CIDs 39252,55

+++ src/postscriptlight.c
@@ -3593,7 +3593,7 @@ int PSL_plotbitimage (struct PSL_CTRL *PSL, double x, double y, double xsize, do
 	/* Correct origin (x,y) in case of justification */
 	if (justify > 1) {      /* Move the new origin so (0,0) is lower left of box */
 		x -= 0.5 * ((justify + 3) % 4) * xsize;
-		y -= 0.5 * (justify / 4) * ysize;
+		y -= 0.5 * (int)(justify / 4) * ysize;
 	}
 
 	PSL_comment (PSL, "Start of 1-bit image\n");
@@ -3999,7 +3999,7 @@ int PSL_plotepsimage (struct PSL_CTRL *PSL, double x, double y, double xsize, do
 	/* Correct origin (x,y) in case of justification */
 	if (justify > 1) {      /* Move the new origin so (0,0) is lower left of box */
 		x -= 0.5 * ((justify + 3) % 4) * xsize;
-		y -= 0.5 * (justify / 4) * ysize;
+		y -= 0.5 * (int)(justify / 4) * ysize;
 	}
 
 	PSL_command (PSL, "PSL_eps_begin\n");
